feat(particle): Support `bloom` for limiter chain.
[Sundial] Add window sliding function: `roll`, `rollDate`.

+++ Milkomeda/src/main/java/com/github/yizzuide/milkomeda/util/DataTypeConvertUtil.java
@@ -90,7 +90,7 @@ public class DataTypeConvertUtil {
      * @return Integer
      */
     public static Integer intVal(Long num) {
-        return num.intValue();
+        return Math.toIntExact(num);
     }
 
     /**
